+++ CHANGELOG.md
@@ -10,6 +10,7 @@
  - Fix CCS811 Input measurement ([#467](https://github.com/kizniche/mycodo/issues/467))
  - Fix pigpio dependency install issue
  - Prevent pre-output from remaining on after an Input is deactivated
+ - Enable unit conversions for AM2315
 
 ### Miscellaneous